requirements.txt
streamlit
PyPDF2
python-docx
sentence-transformers
faiss-cpu
numpy
pandas
tabula-py
transformers
torch
requests
pathlib
